docs(runtime): clarify ingress routing boundary
